Move measurement/update_line to display/
Because it has nothing to do with measurement.
