+++ src/display/update_display.js
@@ -2,12 +2,12 @@ import { sawCollapsedSpans } from "../line/saw_special_spans"
 import { heightAtLine, visualLineEndNo, visualLineNo } from "../line/spans"
 import { getLine, lineNumberFor } from "../line/utils_line"
 import { displayHeight, displayWidth, getDimensions, paddingVert, scrollGap } from "../measurement/position_measurement"
-import { buildLineElement, updateLineForChanges } from "../measurement/update_line"
 import { mac, webkit } from "../util/browser"
 import { activeElt, removeChildren } from "../util/dom"
 import { hasHandler, signal } from "../util/event"
 import { indexOf } from "../util/misc"
 
+import { buildLineElement, updateLineForChanges } from "./update_line"
 import { startWorker } from "./highlight_worker"
 import { maybeUpdateLineNumberWidth } from "./line_numbers"
 import { measureForScrollbars, updateScrollbars } from "./scrollbars"
