+++ src/display/update_line.js
